Regroup PMP named parameters with BGL named parameters
