+++ BGL/include/CGAL/boost/graph/named_function_params.h
@@ -78,6 +78,7 @@ namespace boost{
 
 namespace CGAL {
 namespace internal_np{
+enum all_default_t { all_default }; //cannot use macro because it takes no argument
 
 // for uniformity we import them in this namespace. Note that
 // it is an import so that if we use the named parameter function
@@ -91,6 +92,7 @@ using boost::visitor;
 #define CGAL_add_named_parameter(X, Y, Z)            \
   enum X { Y };
 #include <CGAL/boost/graph/parameters_interface.h>
+#include <CGAL/Polygon_mesh_processing/internal/parameters_interface.h>
 #undef CGAL_add_named_parameter
 
 }//internal_np
@@ -103,6 +105,12 @@ using boost::visitor;
 
     cgal_bgl_named_params(T v = T()) : base(v) {}
     cgal_bgl_named_params(T v, const Base& b) : base(v, b) {}
+    cgal_bgl_named_params<bool, internal_np::all_default_t, self>
+    all_default() const
+    {
+      typedef cgal_bgl_named_params<bool, internal_np::all_default_t, self> Params;
+      return Params(*this);
+    }
 
 // create the functions for new named parameters and the one imported boost
 // used to concatenate several parameters
@@ -116,10 +124,27 @@ using boost::visitor;
   }
 #include <CGAL/boost/graph/parameters_interface.h>
 #include <CGAL/boost/graph/boost_parameters_interface.h>
+#include <CGAL/Polygon_mesh_processing/internal/parameters_interface.h>
 #undef CGAL_add_named_parameter
   };
 
-  namespace parameters {
+namespace parameters {
+
+  cgal_bgl_named_params<bool, internal_np::all_default_t>
+  inline all_default()
+  {
+    typedef cgal_bgl_named_params<bool, internal_np::all_default_t> Params;
+    return Params();
+  }
+
+
+  template <typename T, typename Tag, typename Base>
+  cgal_bgl_named_params<T,Tag,Base>
+  inline no_parameters(cgal_bgl_named_params<T,Tag,Base>)
+  {
+    typedef cgal_bgl_named_params<T,Tag,Base> Params;
+    return Params();
+  }
 
 // define free functions for named parameters
 #define CGAL_add_named_parameter(X, Y, Z)        \
@@ -132,72 +157,11 @@ using boost::visitor;
   }
 #include <CGAL/boost/graph/parameters_interface.h>
 #include <CGAL/boost/graph/boost_parameters_interface.h>
+#include <CGAL/Polygon_mesh_processing/internal/parameters_interface.h>
 #undef CGAL_add_named_parameter
 
   } // namespace parameters
 
-  //helper classes
-  template<typename PolygonMesh, typename PropertyTag>
-  class property_map_selector
-  {
-  public:
-    typedef typename boost::graph_has_property<PolygonMesh, PropertyTag>::type Has_internal_pmap;
-    typedef typename boost::mpl::if_c< Has_internal_pmap::value
-                            , typename boost::property_map<PolygonMesh, PropertyTag>::type
-                            , typename boost::cgal_no_property::type
-    >::type type;
-    typedef typename boost::mpl::if_c< Has_internal_pmap::value
-                            , typename boost::property_map<PolygonMesh, PropertyTag>::const_type
-                            , typename boost::cgal_no_property::const_type
-    >::type const_type;
-
-    type get_pmap(const PropertyTag& p, PolygonMesh& pmesh)
-    {
-      return get_impl(p, pmesh, Has_internal_pmap());
-    }
-
-    const_type get_const_pmap(const PropertyTag& p, const PolygonMesh& pmesh)
-    {
-      return get_const_pmap_impl(p, pmesh, Has_internal_pmap());
-    }
-
-  private:
-    type get_impl(const PropertyTag&, PolygonMesh&, CGAL::Tag_false)
-    {
-      return type(); //boost::cgal_no_property::type
-    }
-    type get_impl(const PropertyTag& p, PolygonMesh& pmesh, CGAL::Tag_true)
-    {
-      return get(p, pmesh);
-    }
-
-    const_type get_const_pmap_impl(const PropertyTag&
-                                 , const PolygonMesh&, CGAL::Tag_false)
-    {
-      return const_type(); //boost::cgal_no_property::type
-    }
-    const_type get_const_pmap_impl(const PropertyTag& p
-                                 , const PolygonMesh& pmesh, CGAL::Tag_true)
-    {
-      return get(p, pmesh);
-    }
-  };
-
-  template<typename PolygonMesh, typename PropertyTag>
-  typename property_map_selector<PolygonMesh, PropertyTag>::type
-  get_property_map(const PropertyTag& p, PolygonMesh& pmesh)
-  {
-    property_map_selector<PolygonMesh, PropertyTag> pms;
-    return pms.get_pmap(p, pmesh);
-  }
-
-  template<typename PolygonMesh, typename PropertyTag>
-  typename property_map_selector<PolygonMesh, PropertyTag>::const_type
-  get_const_property_map(const PropertyTag& p, const PolygonMesh& pmesh)
-  {
-    property_map_selector<PolygonMesh, PropertyTag> pms;
-    return pms.get_const_pmap(p, pmesh);
-  }
 } //namespace CGAL
 
 // partial specializations hate inheritance and we need to repeat
